[#S06P31A301-24] feat: google login - id_token을 받고 JWT token 반환

+++ server/src/main/java/com/darly/api/service/User/UserService.java
@@ -5,5 +5,5 @@ import com.darly.db.entity.User;
 import java.util.Optional;
 
 public interface UserService {
-    Optional<User> getUserByUserId(int userId);
+    Optional<User> getUserByUserId(Long userId);
 }
