+++ server/src/main/java/com/darly/db/repository/UserRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findUserByUserId(Integer userId);
+    Optional<User> findUserByUserId(Long userId);
     Optional<User> findUserByUserEmail(String userEmail);
 }
